requirements.txt
aiohttp==3.8.4
allure_python_commons==2.13.2
Faker==19.12.0
jsonschema==4.20.0
pytest==7.4.2
Requests==2.31.0
